requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.8
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-phonenumber-field==7.3.0
django-rest-knox==4.2.0
djangorestframework==3.15.1
idna==3.7
phonenumbers==8.13.39
pillow==10.3.0
pycparser==2.22
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
